tooling/ML-EIA-PETROLEUM-INTEL/requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
xlrd>=2.0
statsmodels>=0.14
matplotlib>=3.7
python-docx>=0.8
python-dotenv>=1.0
requests>=2.31
pytest>=7.4
